Now it looks for a global language emulation config
If didn't find any file with a title ID to emulate a language, tries to find a globalLocaleConfig.txt to work for any other title.

Example:
My 3DS is Japanese(because it's a exclusive version), but I want ALL my games in English and they're loading in JP, so instead of having a single txt for each title(that take a lot of effor), it uses globalLocaleConfig.txt changing the language for every game. If there's still a title-specific locale, it'll load it instead of global.